+++ packages/docs/src/pages/en/getting-started/release-notes.md
@@ -4,9 +4,9 @@ meta:
   description: Stay up to date with the latest release notes. The migration guides will also help you migrate applications though major releases.
   keywords: migration, releases, upgrading vuetify
 related:
-  - /introduction/long-term-support/
+  - /overview/long-term-support/
   - /getting-started/contributing/
-  - /introduction/roadmap/
+  - /overview/roadmap/
 ---
 
 # Release notes